Refactor constants in Math.h and update Laser constructor
Changed constants in Math.h to use inline constexpr for better linkage. Updated Laser constructor to explicitly cast window width to float for clarity and type safety.

+++ Source/GameClean/Game/Laser.h
@@ -16,7 +16,7 @@ public:
 	Laser(const swaws::Transform& transform, std::shared_ptr<swaws::Model> model) :
 		swaws::Actor{ transform, model } 
 	{
-		length = swaws::GetEngine().GetRenderer().GetWindowWidth();
+		length = (float)swaws::GetEngine().GetRenderer().GetWindowWidth();
 	}
 
 	void Update(float dt) override;
